algorithm/requirements.txt
# GATv2 算法依赖包
# Python 版本要求：3.8

# PyTorch 深度学习框架
torch==2.11.0
torchvision==0.26.0
torchaudio==2.11.0

# PyTorch Geometric 图神经网络库
torch_geometric==2.6.1
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv

# 数据库连接
pymysql==1.0.2

# 数据处理
numpy==1.26.4
scikit-learn==1.3.0

# API 服务
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.4.2

# 日志与工具
python-dotenv==1.0.0
colorlog==6.7.0

# 测试框架
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.24.1
